requirements.txt
streamlit
pandas
plotly
openpyxl
requests
beautifulsoup4
langdetect
python-dateutil
